refactor: 자동 업데이트 후속 개선 — 번들 선택 식별자 우선, 재시작 실패 분리, 모듈 분리
- find_app_bundle_in: DMG 안 여러 .app 중 우리 식별자 번들 우선 (+테스트)
- appApi.autoUpdate는 설치만, 재시작은 스토어가 appApi.restart()로 이어서 요청.
  설치 후 재시작 요청 실패는 UpdateInstalledRestartFailedError로 구분해
  '설치됨 — 다시 실행' 안내(i18n 5개), 릴리즈 노트 예약 유지 (+테스트)
- update_macos.rs(700줄) → update_macos/{mod.rs(순수 함수+테스트), install.rs}
