+++ src/renderer/api/modules/appApi.ts
@@ -76,13 +76,10 @@ export const appApi = {
         new Error('automatic update is only available in the main window'),
       );
     }
-    return runAfterEditorFlush('app update', async () => {
-      const result = await invoke<AppAutoUpdateResult>('app_auto_update', {
-        tag,
-      });
-      await invoke<void>('app_restart');
-      return result;
-    });
+    // 설치만 수행 — 재시작은 호출자가 appApi.restart()로 이어서 요청 (실패를 구분하기 위해 분리)
+    return runAfterEditorFlush('app update', () =>
+      invoke<AppAutoUpdateResult>('app_auto_update', { tag }),
+    );
   },
   // 자동 업데이트 진행 단계 (다운로드 % / 검증 / 설치)
   onUpdateProgress: (
