+++ src/renderer/hooks/app/useUpdateCheck.ts
@@ -1,6 +1,7 @@
 // 이 파일은 useUpdateStore.ts로 이동되었습니다.
 // 기존 import 호환성을 위해 re-export합니다.
 export {
+  UpdateInstalledRestartFailedError,
   useUpdateCheck,
   useUpdateStore,
   hasPendingPostUpdateReleaseNotice,
