+++ src/renderer/locales/zh-cn.json
@@ -461,6 +461,8 @@
     "phaseVerifying": "验证中...",
     "phaseInstalling": "安装中...",
     "phaseRestarting": "重启中...",
+    "phaseInstalled": "已安装 — 请重新启动",
+    "installedRestartFailed": "更新已安装。重新启动 DM NOTE 即可使用新版本。",
     "later": "稍后",
     "checkUpdate": "检查更新",
     "checking": "检查中...",
